change mail views for fileshares and microapps, #8: change mail view for filecollects

+++ app/Http/Controllers/FilecollectController.php
@@ -188,16 +188,16 @@ class FilecollectController extends Controller
                     $extension = ".pdf";
                 }
                 else if($filecollect->fileMime == "application/vnd.openxmlformats-officedocument.wordprocessingml.document"){
-                    $extension = ".doc";
+                    $extension = ".docx";
                 }
 
                 //validate the input file
                 $rule = [
-                    'the_file' => "mimetypes:$filecollect->fileMime"
+                    'the_file' => "file|max:5000|mimetypes:$filecollect->fileMime"
                 ];
                 $validator = Validator::make($request->all(), $rule);
                 if($validator->fails()){ 
-                    return back()->with('failure', 'Μη επιτρεπτός τύπος αρχείου');
+                    return back()->with('failure', $validator->errors()->first());
                 }
 
                 //$file is for the file field of the database
@@ -242,7 +242,7 @@ class FilecollectController extends Controller
             $extension = ".pdf";
         }
         else if($old_data->filecollect->fileMime == "application/vnd.openxmlformats-officedocument.wordprocessingml.document"){
-            $extension = ".doc";
+            $extension = ".docx";
         }
 
         if(get_class($old_data->stakeholder) == 'App\Models\School')
@@ -312,42 +312,46 @@ class FilecollectController extends Controller
     }
 
     public function delete_stakeholder_file(Request $request, FilecollectStakeholder $stakeholder){
-        if(Auth::guard('school')->check()){
-            $identifier = Auth::guard('school')->user()->code;
-        }
-        else if (Auth::guard('teacher')->check()){
-            $identifier = Auth::guard('school')->user()->afm;
-        }
+        if($stakeholder->filecollect->accepts){
+            if(Auth::guard('school')->check()){
+                $identifier = Auth::guard('school')->user()->code;
+            }
+            else if (Auth::guard('teacher')->check()){
+                $identifier = Auth::guard('teacher')->user()->afm;
+            }
 
-        $extension="";
-        if($stakeholder->filecollect->fileMime == "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"){
-            $extension ='.xlsx';
-        }
-        else if($stakeholder->filecollect->fileMime == "application/pdf"){
-            $extension = ".pdf";
-        }
-        else if($stakeholder->filecollect->fileMime == "application/vnd.openxmlformats-officedocument.wordprocessingml.document"){
-            $extension = ".doc";
-        }
+            $extension="";
+            if($stakeholder->filecollect->fileMime == "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"){
+                $extension ='.xlsx';
+            }
+            else if($stakeholder->filecollect->fileMime == "application/pdf"){
+                $extension = ".pdf";
+            }
+            else if($stakeholder->filecollect->fileMime == "application/vnd.openxmlformats-officedocument.wordprocessingml.document"){
+                $extension = ".docx";
+            }
 
-        $directory = "file_collects/$stakeholder->filecollect_id";
-        $original_filename = $identifier.'_filecollect_'.$stakeholder->filecollect_id.$extension;
+            $directory = "file_collects/$stakeholder->filecollect_id";
+            $original_filename = $identifier.'_filecollect_'.$stakeholder->filecollect_id.$extension;
 
-        $fileHandler = New FilesController;
-        try{
-            $fileHandler->delete_file($directory, $original_filename, 'local');
-        }
-        catch(\Exception $e){
-            Log::channel('files')->error($identifier." failed to delete file from filecollect $stakeholder->filecollect_id ".$e->getMessage());
-            return back()->with('failure', 'Το αρχείο δε διαγράφηκε, προσπαθήστε αργότερα ή επικοινωνήστε με τον διαχειριστή του συστήματος');
-        }
-        $stakeholder->uploaded_at = null;
-        $stakeholder->file = null;
-        $stakeholder->checked = null;
-        $stakeholder->save();
+            $fileHandler = New FilesController;
+            try{
+                $fileHandler->delete_file($directory, $original_filename, 'local');
+            }
+            catch(\Exception $e){
+                Log::channel('files')->error($identifier." failed to delete file from filecollect $stakeholder->filecollect_id ".$e->getMessage());
+                return back()->with('failure', 'Το αρχείο δε διαγράφηκε, προσπαθήστε αργότερα ή επικοινωνήστε με τον διαχειριστή του συστήματος');
+            }
+            $stakeholder->uploaded_at = null;
+            $stakeholder->file = null;
+            $stakeholder->checked = null;
+            $stakeholder->stake_comment = null;
+            $stakeholder->save();
 
-        Log::channel('files')->info($identifier." successfully deleted file from filecollect $stakeholder->filecollect_id");
-        return back()->with('success', 'Το αρχείο διαγράφηκε');
+            Log::channel('files')->info($identifier." successfully deleted file from filecollect $stakeholder->filecollect_id");
+            return back()->with('success', 'Το αρχείο διαγράφηκε');
+        }
+        else abort(403);
     }
 
     public function save_filecollect_comment(Request $request, FilecollectStakeholder $stakeholder){
